Added authentication to MERN app

+++ mern/frontend/src/components/TaskDetails.js
@@ -1,14 +1,23 @@
 import { useTaskContext } from "../hooks/useTaskContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 // date fns
 import { formatDistanceToNow, format } from 'date-fns';
 
 const TaskDetails = ({ task }) => {
     const { dispatch } = useTaskContext();
+    const { user } = useAuthContext();
     
     const handleClick = async () => {
+        if (!user) {
+            return;
+        }
+        
         const response = await fetch('api/tasks/' + task._id, {
             method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         });
         const json = await response.json();
